Refactor extra tags
This is an update to the PR for #356

Instead of taking a list of explicit tags, simply take a map, it's easier
to write the spec.

Build up tags in a map and turn that into a string at the end, to avoid
fragility in terminating strings in future updates.