Rewrite allOf support
I believe this fixes #531.

The original way of implementing AllOf was too simple, and what
we need to do is merge the openapi3 definitions into a composite
object before generating code from it.

The code generated is incompatible with that previously generated,
so we're added a config file option and command line flag to revert
to the old behavior, but the new behavior is defaulted.