Update enum conflict resolution.
My last commit missed all the files!

Enums are as short as possible now, and if there are no conflicts,
we use types of the same names of the values, eg, a string enum with
values [One, Two] would generate the Go constants One and Two. However,
if these values would conflict with any global values, we prepend the
typename to them, eg, EnumOne, EnumTwo, etc.

This change introduces an incompatibility, so it can be overridden
to old behavior via the `old-enum-conflicts` config file property or
command line argument.