Generate aliases instead of typedefs
This fixes #579, which turned out to be a big issue, because the
type redefinition which we were doing was hiding things like
JSON Marshaler/Unmarshaler behavior in helper types.

The right fix is to use aliased types - it also makes the code
easier to work with, because less type conversion is needed

This is a breaking change, so to generate code of the old structure,
we can specify the `--old-aliasing` flag.  The `--alias-types` flag
is now ignored.