Rewrite configuration options
The way configuration was handled was becoming a mess. I've
rewritten it to expose the configuration struct from the
codegen package directly.

This is a very breaking change, so the `--old-config-style` flag will
use the old configuration parsing, and `--output-config` will output
the new style config file from old config to aid in rewriting.