Ensure embedded spec contains all necessary OpenAPI documents
When using a schema that uses `$ref`s to an external file, although we
generate the mappings appropriately, this then leads to the embedded
spec being misaligned, as it continues to use the `$ref` to that
external file, which then is not available when calling i.e.
`GetSwagger`.

We can take advantage of the `InternalizeRefs` method to perform an
in-place modification to apply those external references into a single,
embeddable spec.

This appears to only work when a full file is referenced, instead of a
path within a file.