Correctly `go:generate` test client code
It looks like the generation for this has been broken for a bit, and
we've been printing to `stdout`, as no `OutputFile` has been determined.

It appears that this can be fixed by just explicitly adding the
`--generate` flag, but that doesn't quite explain why the `OutputFile`
couldn't be determined.